refactor: move emergence detection into count.py
Merge detect_emergence() and emergence_density_curve() into
count.py per review feedback, since they operate on
count_detections() output and conceptually belong alongside it.
Removes the standalone emergence.py module.
